requirements.txt
altair==5.2.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==23.3.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
ec2-hibinit-agent==1.0.0
Flask==3.0.2
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
kafka-python==2.0.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
packaging==23.2
pandas==2.2.1
pexpect==4.8.0
pillow==10.2.0
protobuf==4.25.3
ps==0.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py4j==0.10.9.7
pyarrow==15.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pyspark==3.5.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
rich==13.7.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sos==4.5.6
SQLAlchemy==2.0.28
ssh-import-id==5.11
streamlit==1.32.0
systemd-python==234
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
Twisted==22.1.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
watchdog==4.0.0
Werkzeug==3.0.1
zipp==1.0.0
zope.interface==5.4.0
